+++ packages/botonic-plugin-ai-agents/src/tools/message.ts
@@ -1,44 +1,6 @@
 import { tool } from '@langchain/core/tools'
 import { z } from 'zod'
 
-export const handoffToHumanAgent = tool(
-  async () => {
-    return null
-  },
-  {
-    name: 'handoffToHumanAgent',
-    schema: z.object({}),
-    description:
-      'Use this tool strictly when the user wants to talk with a human agent.',
-    returnDirect: true,
-  }
-)
-
-export const outOfContext = tool(
-  async () => {
-    return null
-  },
-  {
-    name: 'outOfContext',
-    schema: z.object({}),
-    description:
-      'Use this when the user talks about something that is clearly out of your context. Example: "I want to buy a car" when you are a travel agent.',
-    returnDirect: true,
-  }
-)
-
-export const finishConversation = tool(
-  async () => {
-    return null
-  },
-  {
-    name: 'finishConversation',
-    schema: z.object({}),
-    description: 'Use this when the user wants to end the conversation.',
-    returnDirect: true,
-  }
-)
-
 export const generateMessageWithButtons = tool(
   async (input: { text: string; options: string[] }) => {
     console.log('generateMessageWithButtons', input)
@@ -99,16 +61,7 @@ export const generateCarouselMessage = tool(
   }
 )
 
-export const MANDATORY_TOOLS = [
-  handoffToHumanAgent,
-  outOfContext,
-  finishConversation,
+export const MESSAGE_TOOLS = [
   generateMessageWithButtons,
   generateCarouselMessage,
-]
-
-export const EXIT_TOOLS = [
-  handoffToHumanAgent.name,
-  outOfContext.name,
-  finishConversation.name,
 ]
